# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Marvol Facility Cleaning Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: staffId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create a task assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete an assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /issues/{id}/assign:
    patch:
      operationId: assignIssue
      tags: [issues]
      summary: Assign an issue to a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignIssueRequest"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Issue"
...
  /webhooks/sendgrid/inbound:
    post:
      operationId: receiveSendgridInspectorReply
      tags: [messages]
      summary: Accept an authenticated inspector email reply
      security:
        - SendGridInboundSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendgridInboundReplyRequest"
      responses:
        "200": { description: Reply accepted or idempotently replayed }
        "401": { description: Invalid webhook credential }
        "403": { description: "Invalid sender, token, authentication, or conversation" }
...
components:
  securitySchemes:
    InternalCronSecret:
      type: apiKey
      in: header
      name: X-Internal-Cron-Secret
    SendGridInboundSecret:
      type: apiKey
      in: header
      name: X-SendGrid-Inbound-Secret
  schemas:
...
    SendgridInboundReplyRequest:
      type: object
      required: [envelope, from, text]
      properties:
        envelope:
          type: object
          required: [from, to]
          properties:
            from: { type: string, format: email }
            to: { type: array, items: { type: string, format: email } }
        from: { type: string, format: email }
        text: { type: string, maxLength: 2000 }
        headers: { type: string }
        SPF: { type: string }
        dkim: { type: string }
...
    Assignment:
      type: object
      properties:
        id:
          type: integer
        staffId:
          type: integer
        staffName:
          type: string
        areaId:
          type: integer
        areaName:
          type: string
        terminal:
          type: string
        assignmentDate:
          type: string
        assignedById:
          type: integer
        assignedByName:
          type: string
        notes:
          type: string
          nullable: true
        isSpecial:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - staffId
        - staffName
        - areaId
        - areaName
        - terminal
        - assignmentDate
        - assignedById
        - assignedByName
        - isSpecial
        - createdAt
    CreateAssignmentRequest:
      type: object
      properties:
        staffId:
          type: integer
        areaId:
          type: integer
        assignmentDate:
          type: string
        assignedById:
          type: integer
        notes:
          type: string
          nullable: true
        isSpecial:
          type: boolean
      required:
        - staffId
        - areaId
        - assignmentDate
        - assignedById
        - isSpecial
    Issue:
      type: object
      properties:
        id:
          type: integer
        areaId:
          type: integer
        areaName:
          type: string
        reportedById:
          type: integer
        reportedByName:
          type: string
        reportedByActive:
          type: boolean
          description: Whether the staff member who reported the issue is still active.
        assignedToId:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        assignedToActive:
          type: boolean
          nullable: true
          description: Whether the assigned staff member is still active. Null when no one is assigned.
        issueDate:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        resolved:
          type: boolean
        resolvedAt:
          type: string
          nullable: true
        completionNotes:
          type: string
          nullable: true
        beforeImagePath:
          type: string
          nullable: true
        afterImagePath:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - areaId
        - areaName
        - reportedById
        - reportedByName
        - reportedByActive
        - issueDate
        - description
        - severity
        - resolved
        - createdAt
...
    AssignIssueRequest:
      type: object
      properties:
        assignedToId:
          type: integer
          nullable: true
        assignedById:
          type: integer
      required:
        - assignedById
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...